chore: move 'required_script' in its own module.
